Extract Internals module from Data.Persist
This allows people that want to access the internals of the format to do more
complicated or optimized things, while hiding the internals from most users.
We also expose IO versions of evalPut and runGet. This allows users who are
already in IO to avoid unsafePerformIO and the NOINLINE barrier. More
complicated evaluation strategies, such as using the remainder of some other
application buffer as the first chunk, are now available.