Add unsafe versions of put/get and helpers
Allow for Persist instances to reserve specific sizes ahead of time and
then elide the checks while serializing. As an implemented example, we
allocate space for the length of a ByteString and the bytes in the
ByteString together. For characters as well, we allocate the space all
at once. This technique would be useful for arrays of fixed size
elements, like unboxed vectors.