Add getPrefix and unsafeGetPrefix
It often occurs when deserializing a custom format that we need to
recursively deserialize a subcomponent from the next `n` bytes. The
naive way to do this would be to get a `ByteString` for those n bytes,
and then run the appropriate `Get` on that `ByteString`. To simplify
this process, introduce `getPrefix` and `unsafeGetPrefix`. These
manipulate the `Get` internals to expose only the appropriate bytes to
the sub-deserializer. This avoids repeated copying, and even repeated
`ByteString` construction.