Add length backpatching
There are many serialized formats where a serialized sub-value is
prefixed by its length. The simplest way to find the serialized length
of a value is to serialize it and then to measure the length. For
formats that use a fixed number of bytes for this length, we can avoid
copying and also walking a data structure twice: Once to find the size,
and again to write the data with the known size. It is straightforward
to use this. Reserve the size, serialize the sub-value, then resolve the
size, with the correct endianness and inclusivity.

We do this by reserving the bytes ahead of time and returning a
`PutSize` handle that can be used later to fill in the result.
Currently, the library handles computing the appropriate size, which can
be filled in either Big or Little endian, and either inclusive or
exclusive of the size itself. DHCPv6 packets use this style of
length-value encoding. As does Cassandra's encoding format, and MongoDB
as well. MongoDB uses this same backpatching optimization.