Switch to NoFieldSelectors and OverloadedRecordDot
GHC has supported OverloadedRecordDot and NoFieldSelectors since 9.2
Switch away from prefixed names and instead just use names and record
dot syntax. There are a handful of now-ambiguous updates. They were
resolved using a small lambda and RecordWildCards/NamedFieldPuns. This
is a breaking change, but there are several and we are planning on a new
major release.