Improve interaction with bytestring library
Avoid a copy for a strict ByteString with only one chunk by calling
`reallocBytes` and then building the ByteString directly by recording a
finalizer.

Add support for running a put and producing a lazy ByteString. Turn each
chunk into a strict ByteString with the appropriate finalizer, and then
use `foldlM` to reverse the chunks and build the lazy ByteString at the
same time.