Add resolveSize helpers that write a computed size
This makes list serialization much faster. It makes almost all of the
serialization benchmarks competitive with store. The remaining gap is
due to reinterpretCast. Copying a Double to tmp and then again to the
buffer is slower. But this is a simple win that puts the new reservation
code to good use.